+++ test/features/secrets-cache-feature.js
@@ -3,14 +3,14 @@ import path from 'node:path/posix';
 import { mock } from 'node:test';
 
 import { SecretsCache } from '@aller/google-cloud-secret';
+import { startServer, RpcCodes } from '@aller/google-cloud-secret/fake-server/fake-secret-manager-server';
 import secretManager from '@google-cloud/secret-manager';
+import * as grpc from '@grpc/grpc-js';
 
-import { RpcCodes } from '../../src/fake-server/rpc-codes.js';
 import { fakeAuth } from '../helpers/fake-auth.js';
-import { startServer, reset } from '../helpers/fake-server.js';
 
 Feature('secrets cache', () => {
-  /** @type {import('@grpc/grpc-js').Server} */
+  /** @type {Awaited<ReturnType<typeof startServer>>} */
   let server;
   /** @type {import('@google-cloud/secret-manager').SecretManagerServiceClient} */
   let client;
@@ -18,14 +18,14 @@ Feature('secrets cache', () => {
     server = await startServer();
     client = new secretManager.v1.SecretManagerServiceClient({
       apiEndpoint: 'localhost',
+      sslCreds: grpc.credentials.createInsecure(),
       port: server.origin.port,
       auth: fakeAuth(),
     });
   });
   after(async () => {
-    client = await client.close();
-    server = server?.forceShutdown();
-    reset();
+    await client.close();
+    server?.forceShutdown();
   });
   after(() => mock.timers.reset());
 
@@ -110,6 +110,7 @@ Feature('secrets cache', () => {
     before(() => {
       cache = new SecretsCache({
         apiEndpoint: 'localhost',
+        sslCreds: grpc.credentials.createInsecure(),
         port: server.origin.port,
         auth: fakeAuth(),
       });